Use in-process distributed lock provider in test factories
The OIDC PreAuthSessionStore now uses IDistributedLockProvider (since
3baabebe). In integration tests Redis is disabled, so AddDistributedLocking
falls back to SqlDistributedSynchronizationProvider against the production
connection string (Server=localhost,1433). In CI that address is unreachable,
so OidcPreAuthSecurityTests.{CrossSession_CompleteLogin,V02_T02_CompleteLogin
_SecondCallWithConsumedSession,SessionReplay_CallbackOnAlreadyUsedSession}
hung on the SQL connection timeout (14s) and then failed.

Extracts the existing InProcessLockProvider from PreAuthSessionStoreTests
into a shared Helpers/ location, then registers it via ConfigureServices in
both PortalWebApplicationFactory and PluginIntegrationWebApplicationFactory.
The provider has the same mutual-exclusion semantics as the Redis and SQL
providers (per-key SemaphoreSlim, blocking acquire), so the OIDC tests still
exercise the lock-serialized read-modify-write — just without an external
dependency.
