+++ test/SEBT.Portal.Tests/Integration/PortalWebApplicationFactory.cs
@@ -1,9 +1,11 @@
+using Medallion.Threading;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Mvc.Testing;
 using Microsoft.Extensions.DependencyInjection;
 using NSubstitute;
 using SEBT.Portal.Core.Services;
 using SEBT.Portal.Infrastructure.Services;
+using SEBT.Portal.Tests.Helpers;
 
 namespace SEBT.Portal.Tests.Integration;
 
@@ -77,6 +79,15 @@ public class PortalWebApplicationFactory : WebApplicationFactory<Program>
                 && d.ImplementationType?.FullName?.Contains("Redis", StringComparison.OrdinalIgnoreCase) == true);
             if (redisDescriptor != null)
                 services.Remove(redisDescriptor);
+
+            // Replace the distributed lock provider with an in-process implementation.
+            // Without this, AddDistributedLocking falls back to SqlDistributedSynchronizationProvider
+            // (since Redis is disabled above), which tries to open a real SQL connection on lock
+            // acquisition — failing in CI where no SQL Server is reachable at the default address.
+            var lockDescriptor = services.SingleOrDefault(d => d.ServiceType == typeof(IDistributedLockProvider));
+            if (lockDescriptor != null)
+                services.Remove(lockDescriptor);
+            services.AddSingleton<IDistributedLockProvider>(new InProcessLockProvider());
         });
     }
